# spec/openapi.yaml
openapi: 3.0.3
info:
  title: markdown-manager API
  version: 1.0.0
  description: >
    Ingest markdown/text, chunk it, embed each chunk with a local model, and
    search stored chunks by vector similarity. Embedding runs asynchronously,
    so ingest endpoints return 202 and the resource starts in PENDING status.

servers:
  - url: http://localhost:8000
    description: Local development

tags:
  - name: markdowns
    description: Ingest and inspect markdown documents
  - name: search
    description: Semantic similarity search over chunks

paths:
  /api/markdowns/:
    get:
      tags: [markdowns]
      summary: List markdowns
      parameters:
        - $ref: '#/components/parameters/Page'
      responses:
        '200':
          description: Paginated list of markdowns
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      results:
                        type: array
                        items: { $ref: '#/components/schemas/Markdown' }
    post:
      tags: [markdowns]
      summary: Create a markdown from JSON
      description: Creates the markdown and enqueues embedding. Returns 202.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MarkdownCreate' }
      responses:
        '202':
          description: Accepted; embedding enqueued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Markdown' }
        '400':
          $ref: '#/components/responses/ValidationError'

  /api/markdowns/upload/:
    post:
      tags: [markdowns]
      summary: Upload a .md file
      description: >
        Upload a UTF-8 encoded `.md` file as multipart form data. The filename
        becomes the title. Creates the markdown and enqueues embedding.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: A UTF-8 encoded .md file
      responses:
        '202':
          description: Accepted; embedding enqueued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Markdown' }
        '400':
          description: Missing file, non-.md extension, or non-UTF-8 content
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FieldError' }

  /api/markdowns/{id}/:
    parameters:
      - $ref: '#/components/parameters/MarkdownId'
    get:
      tags: [markdowns]
      summary: Retrieve a markdown
      responses:
        '200':
          description: The markdown, including current status and chunk count
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Markdown' }
        '404':
          $ref: '#/components/responses/NotFound'

  /api/markdowns/{id}/chunks/:
    parameters:
      - $ref: '#/components/parameters/MarkdownId'
      - $ref: '#/components/parameters/Page'
    get:
      tags: [markdowns]
      summary: List a markdown's chunks
      responses:
        '200':
          description: Paginated list of chunks
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      results:
                        type: array
                        items: { $ref: '#/components/schemas/Chunk' }
        '404':
          $ref: '#/components/responses/NotFound'

  /api/search/:
    post:
      tags: [search]
      summary: Semantic search
      description: Embeds the query and returns the top-k most similar chunks.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SearchRequest' }
      responses:
        '200':
          description: Ranked chunks (ascending cosine distance)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SearchResult' }
        '400':
          $ref: '#/components/responses/ValidationError'

components:
  parameters:
    MarkdownId:
      name: id
      in: path
      required: true
      description: Markdown UUID
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      required: false
      description: Page number (page size is 20)
      schema:
        type: integer
        minimum: 1

  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              detail: { type: string }
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/FieldError' }

  schemas:
    Markdown:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        title:
          type: string
        text:
          type: string
        status:
          type: string
          enum: [PENDING, PROCESSING, DONE, FAILED]
          readOnly: true
        error:
          type: string
          readOnly: true
        chunk_count:
          type: integer
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true

    MarkdownCreate:
      type: object
      required: [text]
      properties:
        title:
          type: string
        text:
          type: string

    Chunk:
      type: object
      properties:
        index:
          type: integer
        text:
          type: string

    SearchRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
        top_k:
          type: integer
          default: 5
          minimum: 1
          maximum: 100

    SearchResult:
      type: object
      properties:
        markdown_id:
          type: string
          format: uuid
        index:
          type: integer
        text:
          type: string
        distance:
          type: number
          format: float

    PaginatedResponse:
      type: object
      properties:
        count:
          type: integer
        next:
          type: string
          format: uri
          nullable: true
        previous:
          type: string
          format: uri
          nullable: true
        results:
          type: array
          items: {}

    FieldError:
      type: object
      description: Per-field validation errors keyed by field name.
      additionalProperties:
        type: array
        items:
          type: string
      example:
        file: ["Only .md files are accepted."]